fix: C.LENGTH_UNSET类型转换, TransferListener导入, SmbFile构造参数

+++ app/src/main/java/com/multimediaplayer/player/datasource/FtpDataSource.kt
@@ -51,7 +51,7 @@ class FtpDataSource : BaseDataSource(false) {
         }
 
         transferStarted(dataSpec)
-        return C.LENGTH_UNSET
+        return C.LENGTH_UNSET.toLong()
     }
 
     @Throws(java.io.IOException::class)
